Update Table samples
